+++ dataLinkNode.h
@@ -8,11 +8,10 @@ struct Class_linkNode
     LinkNode *priorNode;
     LinkNode *nextNode;
     int64_t id;
-    uint32_t memSize;
+    uint16_t memSize;
 
     /* virtual operation */
     void (*_contantDinit)(void *contant);
-
     /* object */
     void *contant;
 };
